requirements.txt
slack-bolt>=1.16.0
python-dotenv>=1.0.0
numpy>=1.22.0
sympy>=1.10.1
matplotlib>=3.5.0
